invitro-scan (V 1.1)/requirements.txt
Flask>=2.0
Flask-Cors
qrcode[pil]
Pillow
reportlab
requests
opencv-python
pyzbar
simpleaudio
pandas
numpy
pywebview
pyinstaller